Adding the forecast page. Took the details from the Dublin section for the 7 day forecast . will look to add the container as Dublin an wrap the 7 day forecast for dublin within that and duplicate that for each 6 citites.
found the details on colour for bulma was using black and grey. Wanted to use blue was reading up on bulma site and seen the color pallete details amended site to link and info colours

was nice with the njk nav an footer I just had t amend them ine one place for the whole website

added the spacing helper for bulma py-2 on the footer as it was to large wanted it smaller
Added some text weight to the Dublin heading som e spacing on the box below as was joined looking
added a card-hidden for the title to keep it in centre but same color as background
generally happy with the dublin page now duplicated it for the other cities and amended the details where needed

Need to look up how to keep footer at bottom of page and stretch content?