Made some changes to the 7 day forecast to add a figure element in with an image on each one
Added the settings page. added the main and two columns

Added the bulma radio buttom elements for the temperature units with the checked element
Added checkbox and a save button

I need to add the base jks and make some css changes to stretch main or bring footer to the bottom all the time as footer if mid page on the city pages

Added base njk took awhile I was making it more complicated than it had to be. I was leaving in html and body with njk footer an nav was duplicating. Was actually grand in the end.

This allowed me to add in the site and site main class so that I can add the css structure file to take care of the page sizing so the footer would remain at the bottom an the page was stretched out.
I couldn't find a way to do this with bulma maybe there is but I had to use css for it